Method giveItemRandomly should calculate double chance values.

+++ L2J_Mobius_3.0_Helios/java/org/l2jmobius/gameserver/model/events/AbstractScript.java
@@ -2877,7 +2877,7 @@ public abstract class AbstractScript extends ManagedScript implements IEventTime
 		
 		long minAmountWithBonus = (long) (minAmount * Config.RATE_QUEST_DROP);
 		long maxAmountWithBonus = (long) (maxAmount * Config.RATE_QUEST_DROP);
-		long dropChanceWithBonus = (long) (dropChance * Config.RATE_QUEST_DROP); // TODO separate configs for rate and amount
+		double dropChanceWithBonus = dropChance * Config.RATE_QUEST_DROP; // TODO separate configs for rate and amount
 		if ((npc != null) && Config.CHAMPION_ENABLE && npc.isChampion())
 		{
 			if ((itemId == Inventory.ADENA_ID) || (itemId == Inventory.ANCIENT_ADENA_ID))
@@ -2921,6 +2921,7 @@ public abstract class AbstractScript extends ManagedScript implements IEventTime
 				{
 					playSound(player, QuestSound.ITEMSOUND_QUEST_ITEMGET);
 				}
+				
 				// if there is no limit, return true every time an item is given
 				if (limit <= 0)
 				{
